Skills: joins, grouping, & filtering
An important part of business is planning for the future and ensuring that the business survives changing market conditions. Some businesses do this remarkably well and last for hundreds of years. In this project, I explored data from BusinessFinancing.co.uk on the world's oldest businesses: when were they founded, and which industries do they belong to?